modificacion de los servicios pt2 cultivos

+++ Backend/src/main/java/biocampo/demo/Domain/Model/SaleDetail.java
@@ -1,6 +1,5 @@
 package biocampo.demo.Domain.Model;
 
-import java.math.BigDecimal;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -16,7 +15,7 @@ public class SaleDetail {
     private Sale sale;
     private Product product;
     private int quantity;
-    private BigDecimal subTotal;
+    private Double subTotal;
     //private String paymentMethod;
     /*public enum PaymentMethod {
         PAYPAL, CARD
